+++ src/client/app/components/orders/orderEdit.component.js
@@ -96,8 +96,8 @@ angular.module('orders')
         		};
         		
         		$scope.onChangeValue = function(fieldChange){
-        			var get = parseFloat($scope.updateOrder.get);
-        			var give = parseFloat($scope.updateOrder.give);
+        			var get = parseInt($scope.updateOrder.get);
+        			var give = parseInt($scope.updateOrder.give);
         			var rate = parseFloat($scope.updateOrder.rate);
         			
         			if(rate > 0){
